+++ tests/Unit/CodeParser/MethodCallsParsingTest.php
@@ -2,7 +2,7 @@
 
 namespace JonasPardon\LaravelEventVisualizer\Tests\Unit\CodeParser;
 
-use JonasPardon\LaravelEventVisualizer\Services\CodeParser;
+use JonasPardon\LaravelEventVisualizer\Services\CodeParser\CodeParser;
 use JonasPardon\LaravelEventVisualizer\Services\CodeParser\ValueObjects\ResolvedCall;
 use JonasPardon\LaravelEventVisualizer\Tests\TestCase;
