+++ tests/Unit/CodeParser/PropertyResolverTest.php
@@ -2,7 +2,7 @@
 
 namespace JonasPardon\LaravelEventVisualizer\Tests\Unit\CodeParser;
 
-use JonasPardon\LaravelEventVisualizer\Services\CodeParser;
+use JonasPardon\LaravelEventVisualizer\Services\CodeParser\CodeParser;
 use JonasPardon\LaravelEventVisualizer\Tests\TestCase;
 use PhpParser\Node;
 use PhpParser\NodeFinder;
